py/requirements.txt
pyserial
jupyter
seaborn